Add Roboto fonts and new refactored RegisterPage.
