+++ MyChat/Modules/Registration/RegistrationPresenter.swift
@@ -1,0 +1,8 @@
+//
+//  RegistrationPresenter.swift
+//  MyChat
+//
+//  Created by Сергей Штейман on 30.11.2022.
+//
+
+import Foundation
